moved api functions to src folder instead of in components folder
